pyproject.toml
[tool.poetry]
name = "gcp-cloud-datastore-caching"
version = "0.1.0"
description = ""
authors = ["Bashir Ali"]
package-mode = false

[tool.poetry.dependencies]
python = "3.12"
google-cloud-datastore = "==2.19.0"
python-dotenv = "==1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
